pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "weightiz"
version = "0.1.0"
description = "Weightiz research repository"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "numpy",
  "pandas",
  "pyyaml",
  "pydantic>=2",
  "pytest",
  "scipy",
]

[project.scripts]
weightiz-research = "weightiz.cli.run_research:main"
weightiz-module5 = "weightiz.cli.run_module5:main"
weightiz-module6 = "weightiz.cli.run_module6:main"
weightiz-server-job = "weightiz.cli.run_server_job:main"
weightiz-validate-artifacts = "weightiz.cli.validate_artifacts:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = [
  "--ignore=scripts/run_sweep_auto.py",
  "--ignore=tests/test_quick_run_mode.py",
  "--ignore=tests/test_dq_selection_ranking.py",
  "--ignore=tests/test_sweep_auto_selection.py",
  "--ignore=tests/test_dq_gate.py",
  "--ignore=tests/test_harness_safeguards.py",
  "--ignore=tests/test_module4_institutional.py",
  "--ignore=tests/test_module5_harness_institutional.py",
  "--ignore=tests/test_strategy_factory_acceptance.py",
]
